# TTS/server/openapi.yaml
...
openapi: 3.0.0
info:
  title: Coqui TTS Server
  description: Text-to-Speech server with voice cloning, multi-speaker, and multi-lingual support
  version: 1.0.0
paths:
  "/api/tts":
    post:
      summary: Text-to-Speech Synthesis
      description: Convert text to speech using the currently loaded TTS model
      operationId: tts_synthesis_post
      parameters:
        - name: text
          in: query
          required: false
          schema:
            anyOf:
              - type: string
            description: Text to synthesize
            title: Text
          description: Text to synthesize
        - name: speaker_id
          in: query
          required: false
          schema:
            anyOf:
              - type: string
            description: Speaker ID for multi-speaker models
            title: Speaker Id
          description: Speaker ID for multi-speaker models
        - name: language_id
          in: query
          required: false
          schema:
            anyOf:
              - type: string
            description: Language ID for multilingual models
            title: Language Id
          description: Language ID for multilingual models
        - name: style_wav
          in: query
          required: false
          schema:
            anyOf:
              - type: string
            description: Style wav file or GST tokens
            title: Style Wav
          description: Style wav file or GST tokens
        - name: speaker_wav
          in: query
          required: false
          schema:
            anyOf:
              - type: string
            description: Speaker wav file for voice cloning
            title: Speaker Wav
          description: Speaker wav file for voice cloning
        - name: speaker-id
          in: header
          required: false
          schema:
            anyOf:
              - type: string
            title: Speaker-Id
        - name: language-id
          in: header
          required: false
          schema:
            anyOf:
              - type: string
            title: Language-Id
        - name: text
          in: header
          required: false
          schema:
            anyOf:
              - type: string

            title: Text
        - name: style-wav
          in: header
          required: false
          schema:
            anyOf:
              - type: string

            title: Style-Wav
        - name: speaker-wav
          in: header
          required: false
          schema:
            anyOf:
              - type: string

            title: Speaker-Wav
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: { }
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
    get:
      summary: Text-to-Speech Synthesis
      description: Convert text to speech using the currently loaded TTS model
      operationId: tts_synthesis_get
      parameters:
        - name: text
          in: query
          required: false
          schema:
            anyOf:
              - type: string

            description: Text to synthesize
            title: Text
          description: Text to synthesize
        - name: speaker_id
          in: query
          required: false
          schema:
            anyOf:
              - type: string

            description: Speaker ID for multi-speaker models
            title: Speaker Id
          description: Speaker ID for multi-speaker models
        - name: language_id
          in: query
          required: false
          schema:
            anyOf:
              - type: string

            description: Language ID for multilingual models
            title: Language Id
          description: Language ID for multilingual models
        - name: style_wav
          in: query
          required: false
          schema:
            anyOf:
              - type: string

            description: Style wav file or GST tokens
            title: Style Wav
          description: Style wav file or GST tokens
        - name: speaker_wav
          in: query
          required: false
          schema:
            anyOf:
              - type: string

            description: Speaker wav file for voice cloning
            title: Speaker Wav
          description: Speaker wav file for voice cloning
        - name: speaker-id
          in: header
          required: false
          schema:
            anyOf:
              - type: string

            title: Speaker-Id
        - name: language-id
          in: header
          required: false
          schema:
            anyOf:
              - type: string

            title: Language-Id
        - name: text
          in: header
          required: false
          schema:
            anyOf:
              - type: string

            title: Text
        - name: style-wav
          in: header
          required: false
          schema:
            anyOf:
              - type: string

            title: Style-Wav
        - name: speaker-wav
          in: header
          required: false
          schema:
            anyOf:
              - type: string

            title: Speaker-Wav
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: { }
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
...
  "/process":
    post:
      summary: MaryTTS Compatible Process
      description: MaryTTS-compatible /process endpoint
      operationId: mary_tts_process_post
      parameters:
        - name: INPUT_TEXT
          in: query
          required: false
          schema:
            anyOf:
              - type: string

            description: Text to synthesize
            title: Input Text
          description: Text to synthesize
        - name: VOICE
          in: query
          required: false
          schema:
            anyOf:
              - type: string

            description: Voice/speaker ID
            title: Voice
          description: Voice/speaker ID
        - name: LOCALE
          in: query
          required: false
          schema:
            anyOf:
              - type: string

            description: Language locale (ignored)
            title: Locale
          description: Language locale (ignored)
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: { }
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
    get:
      summary: MaryTTS Compatible Process
      description: MaryTTS-compatible /process endpoint
      operationId: mary_tts_process_get
      parameters:
        - name: INPUT_TEXT
          in: query
          required: false
          schema:
            anyOf:
              - type: string

            description: Text to synthesize
            title: Input Text
          description: Text to synthesize
        - name: VOICE
          in: query
          required: false
          schema:
            anyOf:
              - type: string

            description: Voice/speaker ID
            title: Voice
          description: Voice/speaker ID
        - name: LOCALE
          in: query
          required: false
          schema:
            anyOf:
              - type: string

            description: Language locale (ignored)
            title: Locale
          description: Language locale (ignored)
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: { }
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
  "/v1/audio/speech":
    post:
      summary: OpenAI Compatible Speech API
      description: OpenAI-compatible text-to-speech endpoint
      operationId: openai_tts_v1_audio_speech_post
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/OpenAITTSRequest"
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: { }
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
...
  "/api/v1/tts":
    post:
      summary: TTS Synthesis
      description: Synthesize speech from text using the advanced API
      operationId: api_v1_tts_api_v1_tts_post
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/TTSRequest"
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: { }
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            "$ref": "#/components/schemas/ValidationError"
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    OpenAITTSRequest:
      properties:
        model:
          type: string
          title: Model
          description: Model to use (ignored, uses currently loaded model)
          default: tts-1
        voice:
          type: string
          title: Voice
          description: Voice ID or file path for voice cloning
        input:
          type: string
          minLength: 1
          title: Input
          description: Text to synthesize
        response_format:
          type: string
          pattern: "^(wav|mp3|opus|aac|flac|pcm)$"
          title: Response Format
          description: Audio format
          default: wav
        speed:
          type: number
          maximum: 4
          minimum: 0.25
          title: Speed
          description: Speed of speech
          default: 1
      type: object
      required:
        - voice
        - input
      title: OpenAITTSRequest
    TTSRequest:
      properties:
        text:
          type: string
          maxLength: 1000
          minLength: 1
          title: Text
          description: Text to synthesize
        speaker:
          anyOf:
            - type: string
          title: Speaker
          description: Speaker ID or speaker wav file path for voice cloning
        language:
          anyOf:
            - type: string
          title: Language
          description: Language ID for multilingual models
        speaker_id:
          anyOf:
            - type: string
          title: Speaker Id
          description: Alias for speaker
        language_id:
          anyOf:
            - type: string
          title: Language Id
          description: Alias for language
        format:
          type: string
          pattern: "^(wav|mp3|opus|aac|flac|pcm)$"
          title: Format
          description: Output audio format
          default: wav
      type: object
      required:
        - text
      title: TTSRequest
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
              - type: string
              - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
        - loc
        - msg
        - type
      title: ValidationError
...